+++ tests/async/test_async_chat.py
@@ -17,14 +17,14 @@ async def test_async_multi_replies(async_client):
     conversation_id = f"test_conv_{conftest.random_word()}"
     num_replies = 3
     prediction = await async_client.chat(
-        "Yo what's up?", return_chatlog=True, max_tokens=5, conversation_id=conversation_id
+        "Yo what's up?", return_chat_history=True, max_tokens=5, conversation_id=conversation_id
     )
-    assert prediction.chatlog is not None
+    assert prediction.chat_history is not None
     for _ in range(num_replies):
         prediction = await prediction.respond("oh that's cool", max_tokens=5)
         assert isinstance(prediction.text, str)
         assert isinstance(prediction.conversation_id, str)
-        assert prediction.chatlog is not None
+        assert prediction.chat_history is not None
         assert prediction.meta
         assert prediction.meta["api_version"]
         assert prediction.meta["api_version"]["version"]
@@ -70,7 +70,11 @@ async def test_async_chat_stream(async_client):
 @pytest.mark.asyncio
 async def test_async_chat_with_connectors_stream(async_client):
     res = await async_client.chat(
-        "How deep in the Mariana Trench", temperature=0, stream=True, connectors=[{"id": "web-search"}]
+        "How deep in the Mariana Trench",
+        temperature=0,
+        stream=True,
+        connectors=[{"id": "web-search"}],
+        prompt_truncation="AUTO",
     )
 
     assert isinstance(res, cohere.responses.chat.StreamingChat)