+++ tests/sync/test_chat.py
@@ -40,17 +40,19 @@ class TestChat(unittest.TestCase):
         with self.assertRaises(cohere.CohereError):
             co.chat("Yo what up?", model="NOT_A_VALID_MODEL").text
 
-    def test_return_chatlog(self):
-        prediction = co.chat("Yo what up?", return_chatlog=True, max_tokens=5)
+    def test_return_chat_history(self):
+        prediction = co.chat("Yo what up?", return_chat_history=True, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
-        self.assertIsNotNone(prediction.chatlog)
-        self.assertGreaterEqual(len(prediction.chatlog), len(prediction.text))
+        self.assertIsNotNone(prediction.chat_history)
+        self.assertIsInstance(prediction.chat_history, list)
+        self.assertEqual(len(prediction.chat_history), 2)
+        self.assertIsInstance(prediction.chat_history[0], dict)
 
-    def test_return_chatlog_false(self):
-        prediction = co.chat("Yo what up?", return_chatlog=False, max_tokens=5)
+    def test_return_chat_history_false(self):
+        prediction = co.chat("Yo what up?", return_chat_history=False, max_tokens=5)
         self.assertIsInstance(prediction.text, str)
 
-        assert prediction.chatlog is None
+        assert prediction.chat_history is None
 
     def test_return_prompt(self):
         prediction = co.chat("Yo what up?", return_prompt=True, max_tokens=5)
@@ -142,15 +144,15 @@ class TestChat(unittest.TestCase):
         prediction = co.chat(
             "Who are you?",
             chat_history=[
-                {"user_name": "User", "message": "Hey!"},
-                {"user_name": "Chatbot", "message": "Hey! How can I help you?"},
+                {"role": "User", "message": "Hey!"},
+                {"role": "Chatbot", "message": "Hey! How can I help you?"},
             ],
             return_prompt=True,
-            return_chatlog=True,
+            return_chat_history=True,
             max_tokens=5,
         )
         self.assertIsInstance(prediction.text, str)
-        self.assertIsNotNone(prediction.chatlog)
+        self.assertIsNotNone(prediction.chat_history)
         self.assertIn("User: Hey!", prediction.prompt)
         self.assertIn("Chatbot: Hey! How can I help you?", prediction.prompt)
 
@@ -256,7 +258,9 @@ class TestChat(unittest.TestCase):
         self.assertGreater(len(prediction.documents), 0)
 
     def test_with_connectors(self):
-        prediction = co.chat("How deep in the Mariana Trench", temperature=0, connectors=[{"id": "web-search"}])
+        prediction = co.chat(
+            "How deep in the Mariana Trench", temperature=0, connectors=[{"id": "web-search"}], prompt_truncation="AUTO"
+        )
         self.assertIsInstance(prediction.text, str)
         self.assertIsInstance(prediction.citations, list)
         self.assertGreater(len(prediction.citations), 0)
@@ -297,7 +301,11 @@ class TestChat(unittest.TestCase):
 
     def test_stream_with_connectors(self):
         prediction = co.chat(
-            "How deep in the Mariana Trench", temperature=0, stream=True, connectors=[{"id": "web-search"}]
+            "How deep in the Mariana Trench",
+            temperature=0,
+            stream=True,
+            connectors=[{"id": "web-search"}],
+            prompt_truncation="AUTO",
         )
 
         self.assertIsInstance(prediction, cohere.responses.chat.StreamingChat)